+++ generators/typescript/sdk/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.21.1] - 2024-06-02
+
+- Improvement: Detect `workerd` (Cloudflare) environments in `Runtime.ts`. The `Stream` class which is 
+  used for Server-Sent Events now prefers `TextDecoder` if it is present in the environment, to 
+  work in Cloudflare environments. 
+
 ## [0.21.0] - 2024-06-05
 
 - Feature: The generator now supports `bigint` types.